DocGptBE/requirements.txt
fastapi
uvicorn
PyJWT[crypto]
pydantic-settings